:recycle: Refactor: 예외처리핸들러를 exception에 독립적으로 분리
